Split ORKResult into separate files for each active task result.
`ORKResult` is getting huge, difficult to maintain, and difficult to handle merge conflicts. I’ve separated out all the active task results into a file for each module result type. Additionally, I split out the `ORKQuestionResult` subclasses into a separate file and the `ORKCollectionResult` subclasses into a separate file.
