+++ ResearchKit/ActiveTasks/ORKFileResult.h
@@ -1,0 +1,81 @@
+/*
+ Copyright (c) 2015, Apple Inc. All rights reserved.
+ 
+ Redistribution and use in source and binary forms, with or without modification,
+ are permitted provided that the following conditions are met:
+ 
+ 1.  Redistributions of source code must retain the above copyright notice, this
+ list of conditions and the following disclaimer.
+ 
+ 2.  Redistributions in binary form must reproduce the above copyright notice,
+ this list of conditions and the following disclaimer in the documentation and/or
+ other materials provided with the distribution.
+ 
+ 3.  Neither the name of the copyright holder(s) nor the names of any contributors
+ may be used to endorse or promote products derived from this software without
+ specific prior written permission. No license is granted to the trademarks of
+ the copyright holders even if such marks are included in this software.
+ 
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+ FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+
+#import <ResearchKit/ORKResult.h>
+
+
+NS_ASSUME_NONNULL_BEGIN
+
+/**
+ The `ORKFileResult` class is a result that references the location of a file produced
+ during a task.
+ 
+ A file result is typically generated by the framework as the task proceeds. When the task
+ completes, it may be appropriate to serialize the linked file for transmission
+ to the server.
+ 
+ Active steps typically produce file results when CoreMotion or HealthKit are
+ serialized to disk using a data logger (`ORKDataLogger`). Audio recording also produces a file
+ result.
+ 
+ When you write a custom step, use files to report results only when the data
+ is likely to be too big to hold in memory for the duration of the task. For
+ example, fitness tasks that use sensors can be quite long and can generate
+ a large number of samples. To compensate for the length of the task, you can stream the samples to disk during
+ the task, and return an `ORKFileResult` object in the result hierarchy, usually as a
+ child of an `ORKStepResult` object.
+ */
+ORK_CLASS_AVAILABLE
+@interface ORKFileResult : ORKResult
+
+/**
+ The MIME content type of the result.
+ 
+ For example, `@"application/json"`.
+ */
+@property (nonatomic, copy, nullable) NSString *contentType;
+
+/**
+ The URL of the file produced.
+ 
+ It is the responsibility of the receiver of the result object to delete
+ the file when it is no longer needed.
+ 
+ The file is typically written to the output directory of the
+ task view controller, so it is common to manage the archiving or cleanup
+ of these files by archiving or deleting the entire output directory.
+ */
+@property (nonatomic, copy, nullable) NSURL *fileURL;
+
+@end
+
+NS_ASSUME_NONNULL_END
+
