+++ ResearchKit/ActiveTasks/ORKHolePegTestResult.h
@@ -1,0 +1,132 @@
+/*
+ Copyright (c) 2015, Shazino SAS. All rights reserved.
+ 
+ Redistribution and use in source and binary forms, with or without modification,
+ are permitted provided that the following conditions are met:
+ 
+ 1.  Redistributions of source code must retain the above copyright notice, this
+ list of conditions and the following disclaimer.
+ 
+ 2.  Redistributions in binary form must reproduce the above copyright notice,
+ this list of conditions and the following disclaimer in the documentation and/or
+ other materials provided with the distribution.
+ 
+ 3.  Neither the name of the copyright holder(s) nor the names of any contributors
+ may be used to endorse or promote products derived from this software without
+ specific prior written permission. No license is granted to the trademarks of
+ the copyright holders even if such marks are included in this software.
+ 
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+ FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+
+#import <ResearchKit/ORKResult.h>
+
+
+NS_ASSUME_NONNULL_BEGIN
+
+@class ORKHolePegTestSample;
+
+/**
+ The `ORKHolePegTestResult` class records the results of a hole peg test.
+ 
+ The hole peg test result object records the number of pegs, an array of move samples, the total duration
+ and also various attributes of the hole peg test.
+ 
+ The hole peg test samples are generated by the framework as the task proceeds.
+ When the task completes, it may be appropriate to serialize them for transmission to a server,
+ or to immediately perform analysis on them.
+ */
+ORK_CLASS_AVAILABLE
+@interface ORKHolePegTestResult : ORKResult
+
+/**
+ The hole peg test moving direction.
+ */
+@property (nonatomic, assign) ORKBodySagittal movingDirection;
+
+/**
+ The step is for the dominant hand.
+ */
+@property (nonatomic, assign, getter = isDominantHandTested) BOOL dominantHandTested;
+
+/**
+ The number of pegs to test.
+ */
+@property (nonatomic, assign) NSInteger numberOfPegs;
+
+/**
+ The detection area sensitivity.
+ */
+@property (nonatomic, assign) double threshold;
+
+/**
+ The hole peg test also assesses the rotation capabilities.
+ */
+@property (nonatomic, assign, getter = isRotated) BOOL rotated;
+
+/**
+ The number of succeeded moves (out of `numberOfPegs` possible).
+ */
+@property (nonatomic, assign) NSInteger totalSuccesses;
+
+/**
+ The number of failed moves.
+ */
+@property (nonatomic, assign) NSInteger totalFailures;
+
+/**
+ The total time needed to perform the test step (ie. the sum of all samples time).
+ */
+@property (nonatomic, assign) NSTimeInterval totalTime;
+
+/**
+ The total distance needed to perform the test step (ie. the sum of all samples distance).
+ */
+@property (nonatomic, assign) double totalDistance;
+
+/**
+ An array of collected samples, in which each item is an `ORKHolePegTestSample`
+ object that represents a peg move.
+ */
+@property (nonatomic, copy, nullable) NSArray *samples;
+
+@end
+
+
+/**
+ The `ORKHolePegTestSample` class represents a peg move.
+ 
+ The sample object records the duration, and the move distance.
+ An hole peg test is included in an `ORKHolePegTestResult` object, and is recorded
+ by the step view controller for the corresponding task.
+ 
+ An hole peg test sample is typically generated by the framework as the task proceeds.
+ When the task completes, it may be appropriate to serialize the sample for
+ transmission to a server, or to immediately perform analysis on it.
+ */
+ORK_CLASS_AVAILABLE
+@interface ORKHolePegTestSample : NSObject <NSCopying, NSSecureCoding>
+
+/**
+ The time interval for the peg move.
+ */
+@property (nonatomic, assign) NSTimeInterval time;
+
+/**
+ The peg move distance.
+ */
+@property (nonatomic, assign) double distance;
+
+@end
+
+NS_ASSUME_NONNULL_END
