+++ ResearchKit/ActiveTasks/ORKPSATStepViewController.m
@@ -38,6 +38,7 @@
 #import "ORKVerticalContainerView.h"
 
 #import "ORKActiveStepViewController_Internal.h"
+#import "ORKPSATResult.h"
 #import "ORKPSATStep.h"
 #import "ORKResult.h"
 #import "ORKStepViewController_Internal.h"
