+++ ResearchKit/ActiveTasks/ORKRangeOfMotionStepViewController.m
@@ -30,6 +30,7 @@
 
 
 #import "ORKRangeOfMotionStepViewController.h"
+
 #import "ORKCustomStepView_Internal.h"
 #import "ORKHelpers_Internal.h"
 #import "ORKActiveStepViewController_Internal.h"
@@ -38,6 +39,7 @@
 #import "ORKDeviceMotionRecorder.h"
 #import "ORKActiveStepView.h"
 #import "ORKProgressView.h"
+#import "ORKRangeOfMotionResult.h"
 #import "ORKSkin.h"
