+++ ResearchKit/ActiveTasks/ORKReactionTimeResult.h
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 2016, Darren Levy. All rights reserved.
+ Copyright (c) 2015, Apple Inc. All rights reserved.
  
  Redistribution and use in source and binary forms, with or without modification,
  are permitted provided that the following conditions are met:
@@ -29,27 +29,33 @@
  */
 
 
-#import "ORKShoulderRangeOfMotionStepViewController.h"
-
-#import "ORKRangeOfMotionResult.h"
-#import "ORKStepViewController_Internal.h"
+#import <ResearchKit/ORKResult.h>
 
 
+NS_ASSUME_NONNULL_BEGIN
 
-@implementation ORKShoulderRangeOfMotionStepViewController
+@class ORKFileResult;
 
-#pragma mark - ORKActiveTaskViewController
+/**
+ The `ORKReactionTimeResult` class represents the result of a single successful attempt within an ORKReactionTimeStep.
+ 
+ The `timestamp` property is equal to the value of systemUptime (in NSProcessInfo) when the stimulus occurred.
+ Each entry of motion data in this file contains a time interval which may be directly compared to timestamp in order to determine the elapsed time since the stimulus.
+ 
+ The fileResult property references the motion data recorded from the beginning of the attempt until the threshold acceleration was reached.
+ Using the time taken to reach the threshold acceleration as the reaction time of a participant will yield a rather crude measurement. Rather, you should devise your own method using the data recorded to obtain an accurate approximation of the true reaction time.
+ 
+ A reaction time result is typically generated by the framework as the task proceeds. When the task
+ completes, it may be appropriate to serialize the sample for transmission to a server
+ or to immediately perform analysis on it.
+ */
+ORK_CLASS_AVAILABLE
+@interface ORKReactionTimeResult: ORKResult
 
-- (ORKResult *)result {
-    ORKStepResult *stepResult = [super result];
-    
-    ORKRangeOfMotionResult *result = [[ORKRangeOfMotionResult alloc] initWithIdentifier:self.step.identifier];
-    result.flexed = 90.0 - _flexedAngle;
-    result.extended = result.flexed + _rangeOfMotionAngle;
-    
-    stepResult.results = [self.addedResults arrayByAddingObject:result] ? : @[result];
-    
-    return stepResult;
-}
+@property (nonatomic, assign) NSTimeInterval timestamp;
+
+@property (nonatomic, strong) ORKFileResult *fileResult;
 
 @end
+
+NS_ASSUME_NONNULL_END
