+++ ResearchKit/ActiveTasks/ORKReactionTimeResult.m
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 2016, Darren Levy. All rights reserved.
+ Copyright (c) 2015, Apple Inc. All rights reserved.
  
  Redistribution and use in source and binary forms, with or without modification,
  are permitted provided that the following conditions are met:
@@ -29,27 +29,57 @@
  */
 
 
-#import "ORKShoulderRangeOfMotionStepViewController.h"
+#import "ORKReactionTimeResult.h"
 
-#import "ORKRangeOfMotionResult.h"
-#import "ORKStepViewController_Internal.h"
+#import "ORKFileResult.h"
+
+#import "ORKResult_Private.h"
+#import "ORKHelpers_Internal.h"
 
 
+@implementation ORKReactionTimeResult
 
-@implementation ORKShoulderRangeOfMotionStepViewController
+- (void)encodeWithCoder:(NSCoder *)aCoder {
+    [super encodeWithCoder:aCoder];
+    ORK_ENCODE_DOUBLE(aCoder, timestamp);
+    ORK_ENCODE_OBJ(aCoder, fileResult);
+}
 
-#pragma mark - ORKActiveTaskViewController
+- (instancetype)initWithCoder:(NSCoder *)aDecoder {
+    self = [super initWithCoder:aDecoder];
+    if (self) {
+        ORK_DECODE_DOUBLE(aDecoder, timestamp);
+        ORK_DECODE_OBJ_CLASS(aDecoder, fileResult, ORKFileResult);
+    }
+    return self;
+}
 
-- (ORKResult *)result {
-    ORKStepResult *stepResult = [super result];
++ (BOOL)supportsSecureCoding {
+    return YES;
+}
+
+- (BOOL)isEqual:(id)object {
+    BOOL isParentSame = [super isEqual:object];
     
-    ORKRangeOfMotionResult *result = [[ORKRangeOfMotionResult alloc] initWithIdentifier:self.step.identifier];
-    result.flexed = 90.0 - _flexedAngle;
-    result.extended = result.flexed + _rangeOfMotionAngle;
-    
-    stepResult.results = [self.addedResults arrayByAddingObject:result] ? : @[result];
-    
-    return stepResult;
+    __typeof(self) castObject = object;
+    return (isParentSame &&
+            (self.timestamp == castObject.timestamp) &&
+            ORKEqualObjects(self.fileResult, castObject.fileResult)) ;
+}
+
+- (NSUInteger)hash {
+    return super.hash ^ [NSNumber numberWithDouble:self.timestamp].hash ^ self.fileResult.hash;
+}
+
+- (instancetype)copyWithZone:(NSZone *)zone {
+    ORKReactionTimeResult *result = [super copyWithZone:zone];
+    result.fileResult = [self.fileResult copy];
+    result.timestamp = self.timestamp;
+    return result;
+}
+
+- (NSString *)descriptionWithNumberOfPaddingSpaces:(NSUInteger)numberOfPaddingSpaces {
+    return [NSString stringWithFormat:@"%@; timestamp: %f; fileResult: %@%@", [self descriptionPrefixWithNumberOfPaddingSpaces:numberOfPaddingSpaces], self.timestamp, self.fileResult.description, self.descriptionSuffix];
 }
 
 @end
