+++ ResearchKit/ActiveTasks/ORKReactionTimeViewController.m
@@ -37,6 +37,7 @@
 #import "ORKActiveStepViewController_Internal.h"
 #import "ORKStepViewController_Internal.h"
 
+#import "ORKReactionTimeResult.h"
 #import "ORKReactionTimeStep.h"
 #import "ORKResult.h"
