+++ ResearchKit/ActiveTasks/ORKToneAudiometryResult.h
@@ -1,0 +1,100 @@
+/*
+ Copyright (c) 2015, Shazino SAS. All rights reserved.
+ 
+ Redistribution and use in source and binary forms, with or without modification,
+ are permitted provided that the following conditions are met:
+ 
+ 1.  Redistributions of source code must retain the above copyright notice, this
+ list of conditions and the following disclaimer.
+ 
+ 2.  Redistributions in binary form must reproduce the above copyright notice,
+ this list of conditions and the following disclaimer in the documentation and/or
+ other materials provided with the distribution.
+ 
+ 3.  Neither the name of the copyright holder(s) nor the names of any contributors
+ may be used to endorse or promote products derived from this software without
+ specific prior written permission. No license is granted to the trademarks of
+ the copyright holders even if such marks are included in this software.
+ 
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+ FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+
+#import <ResearchKit/ORKResult.h>
+
+
+NS_ASSUME_NONNULL_BEGIN
+
+/**
+ The `ORKToneAudiometryResult` class records the results of a tone audiometry test.
+ 
+ The audiometry samples are generated by the framework when the task completes.
+ It may be appropriate to serialize them for transmission to a server,
+ or to immediately perform analysis on them.
+ */
+ORK_CLASS_AVAILABLE
+@interface ORKToneAudiometryResult : ORKResult
+
+/**
+ The system wide output volume set by the user during the audiometry test.
+ 
+ A value in the range `0.0` to `1.0`, with `0.0` representing the minimum volume
+ and `1.0` representing the maximum volume.
+ */
+@property (nonatomic, copy, nullable) NSNumber *outputVolume;
+
+/**
+ An array of collected samples, in which each item is an `ORKToneAudiometrySample`
+ object that represents an audiometry sample.
+ */
+@property (nonatomic, copy, nullable) NSArray<ORKToneAudiometrySample *> *samples;
+
+@end
+
+
+/**
+ The `ORKToneAudiometrySample` class represents an audio amplitude associated
+ with a frequency and a channel.
+ 
+ The sample object records the amplitude, the frequency, and the channel for the audio
+ tone being played. A tone audiometry sample is included in an `ORKToneAudiometryResult`
+ object, and is recorded by the step view controller for the corresponding task
+ when a tap is recognized for a given tone.
+ 
+ A tone audiometry sample is typically generated by the framework as the task proceeds.
+ When the task completes, it may be appropriate to serialize the sample for
+ transmission to a server or to immediately perform analysis on it.
+ */
+ORK_CLASS_AVAILABLE
+@interface ORKToneAudiometrySample : NSObject <NSCopying, NSSecureCoding>
+
+/**
+ The frequency value in hertz for the tone associated with the sample.
+ */
+@property (nonatomic, assign) double frequency;
+
+/**
+ 
+ The channel, either left or right, for the tone associated with the sample.
+ */
+@property (nonatomic, assign) ORKAudioChannel channel;
+
+/**
+ The audio signal amplitude.
+ 
+ The minimum audio sample amplitude needed for the participant to recognize the sound (a double value between 0 and 1).
+ */
+@property (nonatomic, assign) double amplitude;
+
+@end
+
+NS_ASSUME_NONNULL_END
