+++ ResearchKit/ActiveTasks/ORKTowerOfHanoiResult.h
@@ -1,0 +1,113 @@
+/*
+ Copyright (c) 2015, James Cox. All rights reserved.
+ 
+ Redistribution and use in source and binary forms, with or without modification,
+ are permitted provided that the following conditions are met:
+ 
+ 1.  Redistributions of source code must retain the above copyright notice, this
+ list of conditions and the following disclaimer.
+ 
+ 2.  Redistributions in binary form must reproduce the above copyright notice,
+ this list of conditions and the following disclaimer in the documentation and/or
+ other materials provided with the distribution.
+ 
+ 3.  Neither the name of the copyright holder(s) nor the names of any contributors
+ may be used to endorse or promote products derived from this software without
+ specific prior written permission. No license is granted to the trademarks of
+ the copyright holders even if such marks are included in this software.
+ 
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+ FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+
+#import <ResearchKit/ORKResult.h>
+
+
+NS_ASSUME_NONNULL_BEGIN
+
+@class ORKTowerOfHanoiMove;
+
+
+/**
+ The `ORKTowerOfHanoiResult` class records the results of a Tower of Hanoi active task.
+ 
+ An `ORKTowerOfHanoiResult` object records an array of `ORKTowerOfHanoiMove` objects (one for each move)
+ and a Boolean value representing whether the puzzle was solved or not.
+ 
+ An `ORKTowerOfHanoiResult` object is typically generated by the framework as the task proceeds. When the task
+ completes, it may be appropriate to serialize it for transmission to a server
+ or to immediately perform analysis on it.
+ */
+ORK_CLASS_AVAILABLE
+@interface ORKTowerOfHanoiResult : ORKResult
+
+/**
+ A Boolean value indicating whether the puzzle was solved.
+ 
+ The value of this property is `YES` when the puzzle was solved and `NO` otherwise.
+ */
+@property (nonatomic, assign) BOOL puzzleWasSolved;
+
+/**
+ An array of moves, in which each item is an `ORKTowerOfHanoiMove` object that represents a move.
+ */
+@property (nonatomic, copy, nullable) NSArray<ORKTowerOfHanoiMove *> *moves;
+
+@end
+
+
+/**
+ The `ORKTowerOfHanoiMove` class represents a single move in a Tower of Hanoi puzzle.
+ 
+ The Tower of Hanoi move object records the indexes of the donor and recipient towers
+ and the time at which the event occurred. A `towerOfHanoiMove` instance is included in
+ an `ORKTowerOfHanoiResult` object and is recorded by the step view controller for the
+ corresponding task when a move is made.
+ 
+ A Tower of Hanoi move is typically generated by the framework as the task proceeds. When the task
+ completes, it may be appropriate to serialize the move for transmission to a server,
+ or to immediately perform analysis on it.
+ */
+ORK_CLASS_AVAILABLE
+@interface ORKTowerOfHanoiMove : NSObject <NSCopying, NSSecureCoding>
+
+/**
+ A relative timestamp indicating the time of the tap event.
+ 
+ The timestamp is relative to the value of `startDate` in the `ORKResult` object that includes this move.
+ The start date of that object represents the time at which the first move was made.
+ */
+@property (nonatomic, assign) NSTimeInterval timestamp;
+
+/**
+ The index of the donor tower in the move.
+ 
+ The Tower of Hanoi puzzle has three towers, and so the value of this property is
+ therefore always 0, 1, or 2. The indexes sequentially represent the towers from left to right when they are laid out
+ horizontally and from top to bottom when they are layed out vertically. The index for a given tower is consistent
+ between changes to and from the horizontal and vertical layouts.
+ */
+@property (nonatomic,assign) NSUInteger donorTowerIndex;
+
+/**
+ The index of the recipient tower in the move.
+ 
+ The Tower of Hanoi puzzle has three towers, and so the value of this property is
+ therefore always 0, 1, or 2. The indexes sequentially represent the towers from left to right when they are laid out
+ horizontally and from top to bottom when they are layed out vertically. The index for a given tower is consistent
+ between changes to and from the horizontal and vertical layouts.
+ */
+@property (nonatomic,assign) NSUInteger recipientTowerIndex;
+
+@end
+
+NS_ASSUME_NONNULL_END
