+++ ResearchKit/Common/ORKResult.h
@@ -141,109 +141,6 @@ ORK_CLASS_AVAILABLE
 
 
 /**
- Values that identify the button that was tapped in a tapping sample.
- */
-typedef NS_ENUM(NSInteger, ORKTappingButtonIdentifier) {
-
-    /// The touch landed outside of the two buttons.
-    ORKTappingButtonIdentifierNone,
-    
-    /// The touch landed in the left button.
-    ORKTappingButtonIdentifierLeft,
-    
-    /// The touch landed in the right button.
-    ORKTappingButtonIdentifierRight
-} ORK_ENUM_AVAILABLE;
-
-/**
- The `ORKTappingSample` class represents a single tap on a button.
- 
- The tapping sample object records the location of the tap, the
- button that was tapped, and the time at which the event occurred. A tapping sample is
- included in an `ORKTappingIntervalResult` object, and is recorded by the
- step view controller for the corresponding task when a tap is
- recognized.
- 
- A tapping sample is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize the sample for transmission to a server,
- or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKTappingSample : NSObject <NSCopying, NSSecureCoding>
-
-/**
- A relative timestamp indicating the time of the tap event.
- 
- The timestamp is relative to the value of `startDate` in the `ORKResult` object that includes this
- sample.
- */
-@property (nonatomic, assign) NSTimeInterval timestamp;
-
-/**
- A duration of the tap event.
- 
- The duration store time interval between touch down and touch release events.
- */
-@property (nonatomic, assign) NSTimeInterval duration;
-
-/** 
- An enumerated value that indicates which button was tapped, if any.
- 
- If the value of this property is `ORKTappingButtonIdentifierNone`, it indicates that the tap
- was near, but not inside, one of the target buttons.
- */
-@property (nonatomic, assign) ORKTappingButtonIdentifier buttonIdentifier;
-
-/**
- The location of the tap within the step's view.
- 
- The location coordinates are relative to a rectangle whose size corresponds to
- the `stepViewSize` in the enclosing `ORKTappingIntervalResult` object.
- */
-@property (nonatomic, assign) CGPoint location;
-
-@end
-
-
-/**
- The `ORKTappingIntervalResult` class records the results of a tapping interval test.
- 
- The tapping interval result object records an array of touch samples (one for each tap) and also the geometry of the
- task at the time it was displayed. You can use the information in the object for reference in interpreting the touch
- samples.
- 
- A tapping interval sample is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize it for transmission to a server,
- or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKTappingIntervalResult : ORKResult
-
-/**
- An array of collected samples, in which each item is an `ORKTappingSample` object that represents a
- tapping event.
- */
-@property (nonatomic, copy, nullable) NSArray<ORKTappingSample *> *samples;
-
-/**
- The size of the bounds of the step view containing the tap targets.
- */
-@property (nonatomic) CGSize stepViewSize;
-
-/**
- The frame of the left button, in points, relative to the step view bounds.
- */
-@property (nonatomic) CGRect buttonRect1;
-
-/**
- The frame of the right button, in points, relative to the step view bounds.
- */
-@property (nonatomic) CGRect buttonRect2;
-
-@end
-
-
-/**
  The `ORKPasscodeResult` class records the results of a passcode step.
  
  The passcode result object contains a boolean indicating whether the passcode was saved or not.
@@ -263,598 +160,17 @@ ORK_CLASS_AVAILABLE
 
 @end
 
-/**
- The `ORKRangeOfMotionResult` class records the results of a range of motion active task.
- 
- An `ORKRangeOfMotionResult` object records the flexion and extension values in degrees.
- */
 
-ORK_CLASS_AVAILABLE
-@interface ORKRangeOfMotionResult : ORKResult
 
-/**
- The degrees when bent.
- */
-@property (nonatomic, assign) double flexed;
 
-/**
- The degrees when extended.
-  */
-@property (nonatomic, assign) double extended;
 
-@end
 
 
-/**
- The `ORKTowerOfHanoiResult` class records the results of a Tower of Hanoi active task.
- 
- An `ORKTowerOfHanoiResult` object records an array of `ORKTowerOfHanoiMove` objects (one for each move)
- and a Boolean value representing whether the puzzle was solved or not.
- 
- An `ORKTowerOfHanoiResult` object is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize it for transmission to a server
- or to immediately perform analysis on it.
- */
-@class ORKTowerOfHanoiMove;
 
-ORK_CLASS_AVAILABLE
-@interface ORKTowerOfHanoiResult : ORKResult
 
-/**
- A Boolean value indicating whether the puzzle was solved.
- 
- The value of this property is `YES` when the puzzle was solved and `NO` otherwise.
- */
-@property (nonatomic, assign) BOOL puzzleWasSolved;
 
-/**
- An array of moves, in which each item is an `ORKTowerOfHanoiMove` object that represents a move.
- */
-@property (nonatomic, copy, nullable) NSArray<ORKTowerOfHanoiMove *> *moves;
 
-@end
 
-
-/**
- The `ORKTowerOfHanoiMove` class represents a single move in a Tower of Hanoi puzzle.
- 
- The Tower of Hanoi move object records the indexes of the donor and recipient towers
- and the time at which the event occurred. A `towerOfHanoiMove` instance is included in
- an `ORKTowerOfHanoiResult` object and is recorded by the step view controller for the
- corresponding task when a move is made.
- 
- A Tower of Hanoi move is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize the move for transmission to a server,
- or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKTowerOfHanoiMove : NSObject <NSCopying, NSSecureCoding>
-
-/**
- A relative timestamp indicating the time of the tap event. 
- 
- The timestamp is relative to the value of `startDate` in the `ORKResult` object that includes this move.
- The start date of that object represents the time at which the first move was made.
- */
-@property (nonatomic, assign) NSTimeInterval timestamp;
-
-/**
- The index of the donor tower in the move.
- 
- The Tower of Hanoi puzzle has three towers, and so the value of this property is
- therefore always 0, 1, or 2. The indexes sequentially represent the towers from left to right when they are laid out
- horizontally and from top to bottom when they are layed out vertically. The index for a given tower is consistent
- between changes to and from the horizontal and vertical layouts.
- */
-@property (nonatomic,assign) NSUInteger donorTowerIndex;
-
-/**
- The index of the recipient tower in the move. 
- 
- The Tower of Hanoi puzzle has three towers, and so the value of this property is
- therefore always 0, 1, or 2. The indexes sequentially represent the towers from left to right when they are laid out
- horizontally and from top to bottom when they are layed out vertically. The index for a given tower is consistent
- between changes to and from the horizontal and vertical layouts.
- */
-@property (nonatomic,assign) NSUInteger recipientTowerIndex;
-
-@end
-
-
-/**
- The `ORKToneAudiometryResult` class records the results of a tone audiometry test.
-
- The audiometry samples are generated by the framework when the task completes.
- It may be appropriate to serialize them for transmission to a server,
- or to immediately perform analysis on them.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKToneAudiometryResult : ORKResult
-
-/**
- The system wide output volume set by the user during the audiometry test.
-
- A value in the range `0.0` to `1.0`, with `0.0` representing the minimum volume
- and `1.0` representing the maximum volume.
- */
-@property (nonatomic, copy, nullable) NSNumber *outputVolume;
-
-/**
- An array of collected samples, in which each item is an `ORKToneAudiometrySample`
- object that represents an audiometry sample.
- */
-@property (nonatomic, copy, nullable) NSArray<ORKToneAudiometrySample *> *samples;
-
-@end
-
-
-/**
- The `ORKToneAudiometrySample` class represents an audio amplitude associated
- with a frequency and a channel.
-
- The sample object records the amplitude, the frequency, and the channel for the audio
- tone being played. A tone audiometry sample is included in an `ORKToneAudiometryResult`
- object, and is recorded by the step view controller for the corresponding task
- when a tap is recognized for a given tone.
-
- A tone audiometry sample is typically generated by the framework as the task proceeds.
- When the task completes, it may be appropriate to serialize the sample for
- transmission to a server or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKToneAudiometrySample : NSObject <NSCopying, NSSecureCoding>
-
-/**
- The frequency value in hertz for the tone associated with the sample.
- */
-@property (nonatomic, assign) double frequency;
-
-/**
- 
- The channel, either left or right, for the tone associated with the sample.
- */
-@property (nonatomic, assign) ORKAudioChannel channel;
-
-/**
- The audio signal amplitude.
-
- The minimum audio sample amplitude needed for the participant to recognize the sound (a double value between 0 and 1).
- */
-@property (nonatomic, assign) double amplitude;
-
-@end
-
-
-/**
- The `ORKSpatialSpanMemoryGameTouchSample` class represents a tap during the
- spatial span memory game.
- 
- A spatial span memory game touch sample is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize it for transmission to a server,
- or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKSpatialSpanMemoryGameTouchSample : NSObject <NSCopying, NSSecureCoding>
-
-/**
- A timestamp (in seconds) from the beginning of the game.
- */
-@property (nonatomic, assign) NSTimeInterval timestamp;
-
-/**
- The index of the target that was tapped.
- 
- Usually, this index is a value that ranges between 0 and the number of targets,
- indicating which target was tapped.
- 
- If the touch was outside all of the targets, the value of this property is -1.
- */
-@property (nonatomic, assign) NSInteger targetIndex;
-
-/**
- A point that records the touch location in the step's view.
- */
-@property (nonatomic, assign) CGPoint location;
-
-/**
- A Boolean value indicating whether the tapped target was the correct one.
- 
- The value of this property is `YES` when the tapped target is the correct
- one, and `NO` otherwise.
- */
-@property (nonatomic, assign, getter=isCorrect) BOOL correct;
-
-@end
-
-
-/// An enumeration of values that describe the status of a round of the spatial span memory game.
-typedef NS_ENUM(NSInteger, ORKSpatialSpanMemoryGameStatus) {
-    
-    /// Unknown status. The game is still in progress or has not started.
-    ORKSpatialSpanMemoryGameStatusUnknown,
-    
-    /// Success. The user has completed the sequence.
-    ORKSpatialSpanMemoryGameStatusSuccess,
-    
-    /// Failure. The user has completed the sequence incorrectly.
-    ORKSpatialSpanMemoryGameStatusFailure,
-    
-    /// Timeout. The game timed out during play.
-    ORKSpatialSpanMemoryGameStatusTimeout
-} ORK_ENUM_AVAILABLE;
-
-/**
- The `ORKSpatialSpanMemoryGameRecord` class records the results of a
- single playable instance of the spatial span memory game.
- 
- A spatial span memory game record is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize it for transmission to a server,
- or to immediately perform analysis on it.
- 
- These records are found in the `records` property of an `ORKSpatialSpanMemoryResult` object.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKSpatialSpanMemoryGameRecord : NSObject <NSCopying, NSSecureCoding>
-
-/**
- An integer used as the seed for the sequence.
- 
- If you pass a specific seed value to another game, you get the same sequence.
- */
-@property (nonatomic, assign) uint32_t seed;
-
-/**
- An array of `NSNumber` objects that represent the sequence that was presented to the user.
- 
- The sequence is an array of length `sequenceLength` that contains a random permutation of integers (0..`gameSize`-1)
- */
-@property (nonatomic, copy, nullable) NSArray<NSNumber *> *sequence;
-
-/**
- The size of the game.
- 
- The game size is the number of targets, such as flowers, in the game.
- */
-@property (nonatomic, assign) NSInteger gameSize;
-
-/**
- An array of `NSValue` objects wrapped in `CGRect` that record the frames of the target
- tiles as displayed, relative to the step view.
- */
-@property (nonatomic, copy, nullable) NSArray<NSValue *> *targetRects;
-
-/**
- An array of `ORKSpatialSpanMemoryGameTouchSample` objects that record the onscreen locations
-the user tapped during the game.
- */
-@property (nonatomic, copy, nullable) NSArray<ORKSpatialSpanMemoryGameTouchSample *> *touchSamples;
-
-/**
- A value indicating whether the user completed the sequence and, if the game was not completed, why not.
- */
-@property (nonatomic, assign) ORKSpatialSpanMemoryGameStatus gameStatus;
-
-/**
- An integer that records the number of points obtained during this game toward
- the total score.
- */
-@property (nonatomic, assign) NSInteger score;
-
-@end
-
-
-/**
- The `ORKSpatialSpanMemoryResult` class represents the result of a spatial span memory step (`ORKSpatialSpanMemoryStep`).
- 
- A spatial span memory result records the score displayed to the user, the number of games, the
- objects recording the actual game, and the user's taps in response
- to the game.
- 
- A spatial span memory result is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize it for transmission to a server,
- or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKSpatialSpanMemoryResult : ORKResult
-
-/**
- The score in the game.
- 
- The score is an integer value that monotonically increases during the game, across multiple rounds.
- */
-@property (nonatomic, assign) NSInteger score;
-
-/**
- The number of games.
- 
- The number of rounds that the user participated in, including successful,
- failed, and timed out rounds.
- */
-@property (nonatomic, assign) NSInteger numberOfGames;
-
-/**
- The number of failures.
- 
- The number of rounds in which the user participated, but did not correctly
- complete the sequence.
- */
-@property (nonatomic, assign) NSInteger numberOfFailures;
-
-/**
- An array that contains the results of the games played.
- 
- Each item in the array is an `ORKSpatialSpanMemoryGameRecord` object.
- */
-@property (nonatomic, copy, nullable) NSArray<ORKSpatialSpanMemoryGameRecord *> *gameRecords;
-
-@end
-
-
-/**
- The `ORKFileResult` class is a result that references the location of a file produced
- during a task.
- 
- A file result is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize the linked file for transmission
- to the server.
- 
- Active steps typically produce file results when CoreMotion or HealthKit are
- serialized to disk using a data logger (`ORKDataLogger`). Audio recording also produces a file
- result.
- 
- When you write a custom step, use files to report results only when the data
- is likely to be too big to hold in memory for the duration of the task. For
- example, fitness tasks that use sensors can be quite long and can generate
- a large number of samples. To compensate for the length of the task, you can stream the samples to disk during
- the task, and return an `ORKFileResult` object in the result hierarchy, usually as a
- child of an `ORKStepResult` object.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKFileResult : ORKResult
-
-/**
- The MIME content type of the result.
- 
- For example, `@"application/json"`.
- */
-@property (nonatomic, copy, nullable) NSString *contentType;
-
-/**
- The URL of the file produced.
- 
- It is the responsibility of the receiver of the result object to delete
- the file when it is no longer needed.
- 
- The file is typically written to the output directory of the
- task view controller, so it is common to manage the archiving or cleanup
- of these files by archiving or deleting the entire output directory.
- */
-@property (nonatomic, copy, nullable) NSURL *fileURL;
-
-@end
-
-
-/**
- The `ORKReactionTimeResult` class represents the result of a single successful attempt within an ORKReactionTimeStep.
- 
- The `timestamp` property is equal to the value of systemUptime (in NSProcessInfo) when the stimulus occurred.
-Each entry of motion data in this file contains a time interval which may be directly compared to timestamp in order to determine the elapsed time since the stimulus.
- 
- The fileResult property references the motion data recorded from the beginning of the attempt until the threshold acceleration was reached.
-Using the time taken to reach the threshold acceleration as the reaction time of a participant will yield a rather crude measurement. Rather, you should devise your own method using the data recorded to obtain an accurate approximation of the true reaction time.
- 
- A reaction time result is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize the sample for transmission to a server
- or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKReactionTimeResult: ORKResult
-
-@property (nonatomic, assign) NSTimeInterval timestamp;
-
-@property (nonatomic, strong) ORKFileResult *fileResult;
-
-@end
-
-
-/**
- The `ORKPSATResult` class records the results of a PSAT test.
- 
- The PSAT result object records the initial digit, an array of addition samples, the total of correct answers
- and also various attributes of the PSAT test.
- 
- The PSAT samples are generated by the framework as the task proceeds.
- When the task completes, it may be appropriate to serialize them for transmission to a server,
- or to immediately perform analysis on them.
- */
-@class ORKPSATSample;
-
-ORK_CLASS_AVAILABLE
-@interface ORKPSATResult : ORKResult
-
-/**
- The PSAT presentation mode.
- */
-@property (nonatomic, assign) ORKPSATPresentationMode presentationMode;
-
-/**
- The time interval between two digits presented.
- PSAT-2" is 2 seconds; PSAT-3" is 3 seconds.
- */
-@property (nonatomic, assign) NSTimeInterval interStimulusInterval;
-
-/**
- The amount of time a digit is shown on the screen (0 second for PASAT).
- */
-@property (nonatomic, assign) NSTimeInterval stimulusDuration;
-
-/**
- The length of the series, that is, the number of additions.
- */
-@property (nonatomic, assign) NSInteger length;
-
-/**
- The number of correct sums given (out of 'length' possible ones).
- */
-@property (nonatomic, assign) NSInteger totalCorrect;
-
-/**
- The number of consecutive correct answers (out of 'length - 1' possible ones).
- Used to overcome the alternate answer strategy.
- */
-@property (nonatomic, assign) NSInteger totalDyad;
-
-/**
- The total time needed to answer all additions (that is, the sum of all the samples times).
- */
-@property (nonatomic, assign) NSTimeInterval totalTime;
-
-/**
- The initial digit.
- */
-@property (nonatomic, assign) NSInteger initialDigit;
-
-/**
- An array of collected samples, in which each item is an `ORKPSATSample`
- object that represents an addition sample.
- */
-@property (nonatomic, copy, nullable) NSArray<ORKPSATSample *> *samples;
-
-@end
-
-
-/**
- The `ORKPSATSample` class represents a numeric answer to an addition question.
- If the answer is correct, the sample object records the presented digit and the numeric answer.
- A PSAT sample is included in an `ORKPSATResult` object and is recorded
- by the step view controller for the corresponding task.
- 
- A PSAT sample is typically generated by the framework as the task proceeds.
- When the task completes, it may be appropriate to serialize the sample for
- transmission to a server, or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKPSATSample : NSObject <NSCopying, NSSecureCoding>
-
-/**
- A Boolean value indicating whether the addition answer was the correct one.
- 
- The value of this property is `YES` when the addition result is the correct
- one, and `NO` otherwise.
- */
-@property (nonatomic, assign, getter=isCorrect) BOOL correct;
-
-/**
- The presented digit.
- */
-@property (nonatomic, assign) NSInteger digit;
-
-/**
- The numeric answer;
- `-1` if no answer is provided.
- */
-@property (nonatomic, assign) NSInteger answer;
-
-/**
- The time interval between the new digit and the answer, or test duration if no answer is provided.
- */
-@property (nonatomic, assign) NSTimeInterval time;
-
-@end
-
-
-/**
- The `ORKHolePegTestResult` class records the results of a hole peg test.
- 
- The hole peg test result object records the number of pegs, an array of move samples, the total duration
- and also various attributes of the hole peg test.
- 
- The hole peg test samples are generated by the framework as the task proceeds.
- When the task completes, it may be appropriate to serialize them for transmission to a server,
- or to immediately perform analysis on them.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKHolePegTestResult : ORKResult
-
-/**
- The hole peg test moving direction.
- */
-@property (nonatomic, assign) ORKBodySagittal movingDirection;
-
-/**
- The step is for the dominant hand.
- */
-@property (nonatomic, assign, getter = isDominantHandTested) BOOL dominantHandTested;
-
-/**
- The number of pegs to test.
- */
-@property (nonatomic, assign) NSInteger numberOfPegs;
-
-/**
- The detection area sensitivity.
- */
-@property (nonatomic, assign) double threshold;
-
-/**
- The hole peg test also assesses the rotation capabilities.
- */
-@property (nonatomic, assign, getter = isRotated) BOOL rotated;
-
-/**
- The number of succeeded moves (out of `numberOfPegs` possible).
- */
-@property (nonatomic, assign) NSInteger totalSuccesses;
-
-/**
- The number of failed moves.
- */
-@property (nonatomic, assign) NSInteger totalFailures;
-
-/**
- The total time needed to perform the test step (ie. the sum of all samples time).
- */
-@property (nonatomic, assign) NSTimeInterval totalTime;
-
-/**
- The total distance needed to perform the test step (ie. the sum of all samples distance).
- */
-@property (nonatomic, assign) double totalDistance;
-
-/**
- An array of collected samples, in which each item is an `ORKHolePegTestSample`
- object that represents a peg move.
- */
-@property (nonatomic, copy, nullable) NSArray *samples;
-
-@end
-
-
-/**
- The `ORKHolePegTestSample` class represents a peg move.
- 
- The sample object records the duration, and the move distance.
- An hole peg test is included in an `ORKHolePegTestResult` object, and is recorded
- by the step view controller for the corresponding task.
- 
- An hole peg test sample is typically generated by the framework as the task proceeds.
- When the task completes, it may be appropriate to serialize the sample for
- transmission to a server, or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKHolePegTestSample : NSObject <NSCopying, NSSecureCoding>
-
-/**
- The time interval for the peg move.
- */
-@property (nonatomic, assign) NSTimeInterval time;
-
-/**
- The peg move distance.
- */
-@property (nonatomic, assign) double distance;
-
-@end
 
 
 /**
@@ -971,32 +287,6 @@ ORK_CLASS_AVAILABLE
 
 /** The answer, or `nil` if the user skipped the question. */
 @property (nonatomic, copy, nullable) NSNumber *booleanAnswer;
-
-@end
-
-
-/**
- The `ORKTimedWalkResult` class records the results of a Timed Walk test.
- 
- The timed walk result object records the duration to complete the trial with a specific distance
- and time limit.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKTimedWalkResult : ORKResult
-/**
- The timed walk distance in meters.
- */
-@property (nonatomic, assign) double distanceInMeters;
-
-/**
- The time limit to complete the trials.
- */
-@property (nonatomic, assign) NSTimeInterval timeLimit;
-
-/**
- The trial duration (that is, the time taken to do the walk).
- */
-@property (nonatomic, assign) NSTimeInterval duration;
 
 @end
 
@@ -1410,65 +700,6 @@ ORK_CLASS_AVAILABLE
  The bezier path components used to create the signature image.
  */
 @property (nonatomic, copy, nullable) NSArray <UIBezierPath *> *signaturePath;
-
-@end
-
-
-/**
- The `ORKTrailmakingTap` class represents a single tap in a trail making test.
- 
- The Trailmaking tap move object records the indexes of tap, if it was in error or not
- and the time at which the event occurred. A `trailmakingTap` instance is included in
- an `ORKTrailmakingResult` object and is recorded by the step view controller for the
- corresponding task when a tap is made.
- 
- A trail making tap is typically generated by the framework as the task proceeds. When the task
- completes, it may be appropriate to serialize the move for transmission to a server,
- or to immediately perform analysis on it.
- */
-ORK_CLASS_AVAILABLE
-@interface ORKTrailmakingTap : NSObject <NSCopying, NSSecureCoding>
-
-/**
- A relative timestamp indicating the time of the tap event.
- 
- The timestamp is relative to the value of `startDate` in the `ORKResult` object that includes this move.
- The start date of that object represents the time at which the first move was made.
- */
-@property (nonatomic, assign) NSTimeInterval timestamp;
-
-/**
- The index of the button tapped.
- */
-@property (nonatomic,assign) NSUInteger index;
-
-/**
- If the button was tapped in error.
- */
-@property (nonatomic, assign) BOOL incorrect;
-
-@end
-
-
-/**
- The `ORKTrailmakingResult` class represents the result of a signature step (`ORKTrailmakingStep`).
- 
- 
- A trail making result is produced by the task view controller when it presents a trail making step.
- 
- */
-ORK_CLASS_AVAILABLE
-@interface ORKTrailmakingResult : ORKResult
-
-/** 
- An array of all taps completed during the test
- */
-@property (nonatomic, copy) NSArray <ORKTrailmakingTap *> *taps;
-
-/** 
- The number of errors generated during the test
- */
-@property (nonatomic) NSUInteger numberOfErrors;
 
 @end
